+++ admin/locker/update_locker.php
@@ -27,8 +27,30 @@ if(isset($_SESSION['admin_id'])){
                         // insert
                         if($lockerObj->insert($locker_details['locker_subscription_id'],$_POST['locker_UID'])){
                             // delete the old locker_UID
+                            if(!$locker_user_details =$lockerObj->fetch_user_locker_details($_POST['locker_UID'])) {
+                                echo '0';
+                                return ;
+                            }
                             if($lockerObj->delete($locker_details['locker_id'])){
                                 echo '1';
+
+                                require_once('../../classes/admins.class.php');
+                                require_once('../../classes/notifications.class.php');
+                                $adminObj = new admins();
+                                $notificationObj = new notifications();
+                                if($admin_id_data = $adminObj->fetch_admin_id_of_admins()){
+
+                                    
+                                    
+                                    foreach ($admin_id_data as $key => $value) {
+                                        
+                                        $notification_info ='Staff '.$_SESSION['admin_user_lastname'].', '.$_SESSION['admin_user_firstname'].' '.$_SESSION['admin_user_middlename'].' modified the locker of ('.$locker_user_details['user_name'].') '.$locker_user_details['user_fullname'].'.';
+                                        
+                                        if(!$notificationObj->insert($_SESSION['admin_user_id'],$value['user_id'],'Logs','logs.png', $notification_info)){
+                                            exit('notification insert error');
+                                        }
+                                    }
+                                }
                             }else{
                                 echo '0';
                             }
