+++ admin/maintenance/delete_maintenance.php
@@ -19,7 +19,29 @@ if(isset($_SESSION['admin_id'])){
             if(isset($_POST['equipment_id']) && intval($_POST['equipment_id'])>0){
                 require_once '../../classes/equipments.class.php';
                 $equipmentsObj = new equipments();
+                if(!$equipments_data = $equipmentsObj->fetch_with_id($_POST['equipment_id'])){
+                    echo '0';
+                    return;
+                }
                 if($equipmentsObj->delete($_POST['equipment_id'])){
+                    if($_SESSION['admin_user_type_details'] != 'admin'){
+                        require_once('../../classes/admins.class.php');
+                        require_once('../../classes/notifications.class.php');
+                        
+                        $adminObj = new admins();
+                        $notificationObj = new notifications();
+                        if($admin_id_data = $adminObj->fetch_admin_id_of_admins()){
+                            foreach ($admin_id_data as $key => $value) {
+                                
+                                $notification_info ='Staff '.$_SESSION['admin_user_lastname'].', '.$_SESSION['admin_user_firstname'].' '.$_SESSION['admin_user_middlename'].' deleted the equipment ('.$equipments_data['equipment_name'].').';
+                                
+                                if(!$notificationObj->insert($_SESSION['admin_user_id'],$value['user_id'],'Logs','logs.png', $notification_info)){
+                                    exit('notification insert error');
+                                }
+                            }
+                        }
+                        
+                    }
                     echo '1';
                 }else{
                     echo '0';
